+++ packages/rho.async/R/05-interfaces.R
@@ -3,6 +3,16 @@ RhoStreamLike <- s7contract::new_interface(
   "RhoStreamLike",
   generics = list(rho_stream_next = rho_stream_next)
 )
+RhoDuplexChannel <- s7contract::new_interface(
+  "RhoDuplexChannel",
+  generics = list(
+    rho_stream_next = rho_stream_next,
+    rho_duplex_send = s7contract::interface_requirement(
+      rho_duplex_send,
+      returns = RhoTask
+    )
+  )
+)
 RhoTaskQueue <- s7contract::new_interface(
   "RhoTaskQueue",
   generics = list(rho_enqueue = rho_enqueue)